python_engine/requirements.txt
numpy
pandas
aiohttp
aiofiles
requests
networkx
matplotlib
scikit-learn
python-louvain
fastapi
uvicorn
python-dotenv
pydantic
groq